Improve /godmode to also heal to full. Add the moon phase name to /time. Address edge bug in /fashion. Fix bugs in /il and /sl chat commands.
Fix bug involving leftover invite records when joining an EntryBoard.
Fix bug when selling the materials gained from recycling equipment.
Fix BBM mangling inventory sort data.
Fix null reference edge case in InstanceGatheringItemManager::EmptySpots.
Fix key doors not properly sharing their state with party members.
Fix missing profile information when examining potential pawn hires.
Fix issue where certain pawn data handlers didn't properly return all the sub-notices they should.

When embodying BBM items that can stack (in unusual circumstances), attempt to merge stacks when possible.
Identify Seabell as the Extreme Mission Group 3.
Move the base stats from CharacterManager to CharacterCommon so they're a bit more convenient to access for CData conversions.

Add Megado Craft Room to safe stages.

Fix missing furniture crafting recipes that would normally be unlocked by achievements. Fix mis-attributed node for hunt board quests.
Expose the training room monsters as a new asset.
Fix missing Num for slayer stone crafting recipes.
Fix illegal items being handed out as part of S3 area supply rewards.

Cache the last LobbyLobbyDataMsg and gather it alongside contexts when you move into a hub to prevent players who were doing emotes from being invisible.
Don't allow writing the transient PawnState.Party to the DB; always write PawnState.None instead.
Make sure hub memory is wiped when changing game modes, to prevent ghost characters from standing around.
Be more consistent about only providing the clan short namein certain spots?

+++ Arrowgene.Ddon.GameServer/Characters/AreaRankManager.cs
@@ -46,7 +46,7 @@ namespace Arrowgene.Ddon.GameServer.Characters
             return [.. pointSupplies.SupplyItemList.Select((x, i) => new CDataRewardItemInfo()
             {
                 Index = (uint)i,
-                ItemId = x.ItemId,
+                ItemId = (uint) x.ItemId,
                 Num = (byte)(x.ItemNum * rewardMult),
             })];
         }
